fix: attached files number not updating
refactor: move file action handler to files button component
